// src/app/model/currentuser.ts
export class CurrentUser {
  
    constructor() { }
    token?: string;
    roles?: string;
    username?: string;
    email_address?: string;
    telephone_number?: string;
    persal_number?: number;
    appointment_number?: string;
    identity_number?: number;
    surname?: string;
    first_name?: string;
    initials?: string;
    title?: number;
    gender?: number;
    post_status_code?: number;
    nature_of_appointment?: number;
    job_title_code?: number;
    job_title_description?: string;
    workweek_code?: number;
    acting_reference?: string;
    organisation_code?: string;
    component_code?: number;
    sub_component_code?: number;
    region_code?: number;
    responsibility_code?: number;
    objective_code?: number;
    occupational_classification?: string;
    occupational_group?: number;
    paypoint_code?: number;
    post_salary_level_code?: number;
    organisation_description?: string;
    responsibility_description?: string;
    region_description?: string;
    objective_description?: string;
    component_description?: string;
    sub_component_description?: string;
    paypoint_description?: string;

    post_salary_level?: number;
    system_approver?: number;
    system_recommender?: number;

 }